Final Update for a few months
I attempted to stabilize this project as much as possible so I wouldn't have to update it for a while. Here's what's different:

- assets folder has been changed to be more consistent with my other websites

- background has been updated to appear more real

- The favicon not showing up issue has been fixed

And overall this should work out-of-the-box and you shouldn't face any major issues with it. If you do, submit a pull request or contact me explaining what the issue is.